autoweb/requirements.txt
amqp==5.2.0
APScheduler==3.6.3
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
cachetools==4.2.2
celery==5.3.6
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.8
defusedxml==0.7.1
Django==5.0.6
django-allauth==0.60.1
django-appconf==1.0.6
django-bootstrap5==23.3
django-cachalot==2.6.2
django-filter==23.5
django-jquery==3.1.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-recaptcha==4.0.0
django-redis==5.4.0
django-select2==8.1.2
docopt==0.6.2
gunicorn==21.2.0
idna==3.7
iniconfig==2.0.0
kombu==5.3.7
num2words==0.5.12
oauthlib==3.2.2
packaging==24.1
phonenumbers==8.13.39
pillow==10.3.0
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
py==1.11.0
pycparser==2.22
PyJWT==2.8.0
pytest==7.1.3
pytest-django==4.8.0
pytest-lazy-fixture==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==13.7
python3-openid==3.2.0
pytz==2024.1
redis==5.0.6
reportlab==4.2.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
tornado==6.4.1
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.15
vine==5.1.0
wcwidth==0.2.13
